refactor: move component files to folders
[Ticket: X]
