+++ src/components/atoms/title.tsx
@@ -1,0 +1,5 @@
+const Title = () => {
+  return <h1 className="text-4xl text-primary-100">Memory Game</h1>;
+};
+
+export { Title };
