--- src/components/game-board.tsx
@@ -1,56 +1,0 @@
-const GameBoard = () => {
-  return (
-    <div className="grid grid-cols-4 items-center justify-items-center gap-2 bg-black p-4">
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-      <p className="grid aspect-square w-16 place-content-center rounded-full bg-primary-300">
-        🤭
-      </p>
-    </div>
-  );
-};
-
-export { GameBoard };
