+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './game-board';
-export * from './score-board';
-export * from './setup';
-export * from './title';
+export * from './atoms';
+export * from './layout';
+export * from './molecules';
+export * from './organisms';
