+++ src/components/organisms/game-board.tsx
@@ -1,0 +1,56 @@
+const GameBoard = () => {
+  return (
+    <section className="grid grid-cols-4 items-center justify-items-center gap-4 p-4">
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+      <p className="grid aspect-square w-28 place-content-center rounded-full bg-primary-300">
+        🤭
+      </p>
+    </section>
+  );
+};
+
+export { GameBoard };
