--- src/components/title.tsx
@@ -1,5 +1,0 @@
-const Title = () => {
-  return <h1 className="text-yellow-900">Memory Game</h1>;
-};
-
-export { Title };
